+++ src/fst/parsex.py
@@ -1335,13 +1335,13 @@ def parse__Assign_targets(src: str, parse_params: Mapping[str, Any] = {}) -> AST
     src_ = '\\' + src if src.startswith('\n') else src  # we allow an initial non-line-continued newline because that is a slice indicator to start on new line and will have line continuation added in slice put to Assign
 
     try:
-        ast = _ast_parse1(f'_=\\\n{src_}\\\n_', parse_params, Assign)  # try assuming src has trailing equals
+        ast = _ast_parse1(f'_=\\\n{src_}\\\n=_', parse_params, Assign)  # check assuming src does not have trailing equals, do this first to catch naked tuple with trailing comma
 
-    except SyntaxError as bare_exc:
+    except SyntaxError as posteq_exc:
         try:
-            ast = _ast_parse1(f'_=\\\n{src_}\\\n=_', parse_params, Assign)  # now check assuming src does not have trailing equals
+            ast = _ast_parse1(f'_=\\\n{src_}\\\n_', parse_params, Assign)  # now try assuming src has trailing equals
 
-        except SyntaxError as posteq_exc:
+        except SyntaxError as bare_exc:
             if _syntax_error_in_loc(bare_exc, src):
                 raise bare_exc from None